repair again

+++ src/fact.cpp
@@ -3,7 +3,7 @@
 long long MyMath::factorial(long long n) {
     if (n < 0)
         return 0;
-    long product = 1;
+    long long product = 1;
     for (long long i = n; i > 0; i--) {
         product *= i;
     }
